code cleanup, removed SamplingPointDataClusters module
